Pull out the git log options as an option.
In my environment '--graph' is a kiss of death. It results in seeing
nothing but graph lines due to the number of branches in flight.

Secondly the colors baked into this break with my color scheme
which has a background of black rending the date unreadable.

So to solve this I need to be able to specify my own options,
so add a new options g:fzf_commits_log_options to allow this.