Nanoseconds instead of seconds for buffer timestmaps
Fzf :Buffers uses a dictionary where it stores the last time accessed of
the current buffers. This is used to sort the buffers.
The timestamp is retrieved from a call to localtime() which returns the
current second the buffer is accessed. However, fzf functions in such a
way that before it enters the newly selected buffer, it enters the
current buffer. Sometimes this causes the last two buffers to have the
same timestamp. Repeatedly switching between two buffers yields
inconsistent results.
This commit fixes the issue by using seconds and microseconds from the
bash call `date`. The buffers are now guaranteed to be sorted by last
accessed.