[fzf#vim#gitfiles] Add arg for changing root
This change introduces a new command `:GFilesCurrentDir` and a new
argument for `fzf#vim#gitfiles`.

The goal of this change is to make working in a mono repo easier.
`:GFilesCurrentDir` works the same as `:GFiles` except that it uses
`getcwd()` as the `dir` instead of the git project root for the `dir` to
the fzf call.

The end result is that `:GFilesCurrentDir` works the same in a project
directory of a mono repo as `:GFiles` would in a regular git project