[Snippets] Fix UltiSnips not being able to enter insert mode
UltiSnips snippets have so called tabstops or placeholders that are to
be filled in by the user, so after expanding a snippet the cursor is
moved to the first such tabstop and vim enters insert mode (or, if
there's a default value, select mode).

But this doesn't work with fzf's :Snippets, because UltiSnips uses
:startinsert which doesn't work from :normal. Fix this by calling
UltiSnips#ExpandSnippet directly (unfortunately needs a virtualedit hack
to position the cursor after the snippet name).