[Snippets] Fix when there is word under cursor
UltiSnips#SnippetsInCurrentScope uses the word under cursor, if any, to
filter candidate snippets. To make it usable, however, fzf's :Snippets
must replace the word instead of appending to it, otherwise the
subsequent expansion fails.

Also, use 'i' instead of 'a', to make mappings such as

    inoremap <silent> <C-_> <C-\><C-O>:Snippets<CR>

feel natural. This will break workflow for people who used to manually
exit insert mode in the middle of a line and then invoke :Snippets, but
I'd hazard a guess there aren't many of them as this workflow was
already somewhat broken by the first bug.