[Snippets] Fix UltiSnips not being able to enter select mode
Turns out calling UltiSnips#ExpandSnippet makes it think vim is in
normal mode, so it doesn't feedkeys("\<Esc>"), and that breaks entering
the select mode as the keys are typed instead.

Maybe it's time to start considering that UltiSnips may be to blame and
see if this can be improved there. Here's my first inquiry:
https://github.com/SirVer/ultisnips/commit/1ca82f76f7a4e7e9beac77f74bda99457bbc8286#r44803007