[Snippets] Fix glitches after UltiSnips snippet expansion
UltiSnips adds a CursorMovedI autocmd that is meant to sync its state
with vim, and this is necessary for some of the features like
placeholders and mirrors and so on. We insert some text in
s:inject_snippet but this doesn't trigger the autocmd, and UltiSnips may
get confused afterwards. To fix this, we need to invoke
UltiSnips#CursorMoved after inserting the text, before invoking
UltiSnips#ExpandSnippet.

This glitch can be reproduced as follows:

1. :Snippets, select APACHE
2. C-J to select the author name placeholder, type "xxx"
3. Esc to normal, G to end, :Snippets, select APACHE
4. C-J to select the author name placeholder, type "xxx"
5. glitch.