[Snippets] Revert the startinsert/feedkeys fix for select mode
Using startinsert and feedkeys feels wrong and brittle. So now that vim
8.2.1978+ has <Cmd> that can be used instead of <C-\><C-O> in insert
mode bindings, we can go back to calling the functions explicitly.

(This should probably be documented somewhere, but it's unlikely to
affect many users, as mapping :Snippets in insert mode never worked well
anyway and few people complained.)