bin/preview.sh: Print content based on context position
bat / batcat supports a `line-range` parameter to specify the range of
the file to display. Set the start position based on the provided CENTER
variable and FZF_DEFAULT_LINES. The preview window will contain only the
lines around the CENTER, not the beginning of the file.

For the default awk-based approach, only print the lines that are
greater than the start point.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@gmail.com>