Add support for sapling SCM
This adds the ability for GFiles to return a list of files from
sapling (https://sapling-scm.org), rather than from git.

This extends git support, rather than adding a separate API for it
because the two can interoperate and it comes handy not to have to think
about the repository type of your project.